+++ application/state/useSystemManager.ts
@@ -4,7 +4,7 @@ import type { I18nContextValue } from '../i18n/I18nProvider';
 import { sessionCapabilitiesStore } from './sessionCapabilitiesStore';
 import type { SessionCapabilities } from '../../domain/systemManager/types';
 import type { useSystemManagerBackend } from './useSystemManagerBackend';
-import { nextPollData } from '../../components/systemManager/listStable';
+import { nextPollData } from '../../domain/systemManager/pollListStable';
 
 type Backend = ReturnType<typeof useSystemManagerBackend>;
